+++ test/Elements/New-TimeElement.Tests.ps1
@@ -1,0 +1,29 @@
+<#
+.SYNOPSIS
+	Tests the features of the `New-TimeElement` cmdlet.
+#>
+Describe "New-TimeElement" {
+	BeforeAll {
+		Import-Module "$PSScriptRoot/../../Html.psd1"
+	}
+
+	It 'should support [DateOnly] objects for the "datetime" attribute' {
+		$value = [DateOnly]::FromDateTime((Get-Date -Year 1974 -Month 5 -Day 3))
+		time -datetime $value | Should -BeExactly '<time datetime="1974-05-03"></time>'
+	}
+
+	It 'should support [DateTime] objects for the "datetime" attribute' {
+		$value = Get-Date -Year 1974 -Month 5 -Day 3 -Hour 8 -Minute 16 -Second 32
+		time -datetime $value | Should -BeLikeExactly '<time datetime="1974-05-03T08:16:32.*"></time>'
+	}
+
+	It 'should support [TimeOnly] objects for the "datetime" attribute' {
+		$value = [TimeOnly]::FromDateTime((Get-Date -Hour 8 -Minute 16 -Second 32))
+		time -datetime $value | Should -BeLikeExactly '<time datetime="08:16:32.*"></time>'
+	}
+
+	# It 'should support [TimeSpan] objects for the "datetime" attribute' {
+	# 	$value = New-TimeSpan
+	# 	time -datetime $value | Should -BeExactly '<time datetime="TODO"></time>'
+	# }
+}