+++ ChangeLog.md
@@ -2,8 +2,8 @@
 
 ## Version [0.6.0](https://github.com/cedx/html.net/compare/v0.5.0...v0.6.0)
 - Added the `Html` prefix to the cmdlet nouns.
-- Added the `New-TrackElement` cmdlet.
-- Added the `track` alias.
+- Added the `New-TimeElement` and `New-TrackElement` cmdlets.
+- Added the `time` and `track` aliases.
 
 ## Version [0.5.0](https://github.com/cedx/html.net/compare/v0.4.0...v0.5.0)
 - Added the `New-ColGroupElement`, `New-DataElement`, `New-StyleElement` and `New-VideoElement` cmdlets.